Gateway for retrieveAverageVisitsPerDayByTrustId
